Migrate to multiworker
